+++ src/util/mod.rs
@@ -11,13 +11,13 @@ pub type Result<T> = ::std::result::Result<T, Box<dyn Error>>;
 
 /// Returns the requested resource if available
 /// If it is not available, loops every 5s and tries again until it succeeds
-pub fn get_retry(url: &str, retry_message: &str) -> reqwest::Response {
+pub fn get_retry(url: &str) -> reqwest::Response {
     let mut result = reqwest::get(url);
 
     while let Err(err) = result {
         error!("{}", err.to_string());
-        println!("{}", retry_message);
-        debug!("An error occurred getting {}, retrying...", url);
+        debug!("Retrying {} ...", url);
+        println!("No internet");
 
         thread::sleep(Duration::from_secs(config::RETRY_TIMEOUT));
         result = reqwest::get(url);